WorldClock: Replace mChangingDateTime with mChangingClockInfo
During "change" modes, preserve the ZonedDateTime configured by the user when
transfering from the Controller to the Presenter. This allows the user to to scroll into
non-canonical dates like 2/30 and 2/31 during the "change" modes. The dates are
normalized when the Mode button is LongPressed. Previously, the
changingDateTime was converted to an epochSeconds, then converted to a
ZonedDateTime, then transfered to the Presenter. This caused problems with
non-canonical dates, because 2/29 would scroll into 3/1, even when the display
was blinking.